moving import_json method to BaseImporter
